Refactoring alertActionHandler context
