+++ ui/components/app/confirmations/alerts/alert-modal/alert-modal.tsx
@@ -32,7 +32,7 @@ import {
 import { useI18nContext } from '../../../../../hooks/useI18nContext';
 import useAlerts from '../../../../../hooks/useAlerts';
 import { Alert } from '../../../../../ducks/confirm-alerts/confirm-alerts';
-import { useAlertActionHandler } from '../../../../../hooks/useAlertActionHandler';
+import { useAlertActionHandler } from '../../../../../contexts/alertActionHandler';
 
 export type AlertModalProps = {
   /**
